📝 자유게시판 리스트 페이지 코드 리펙토링
