+++ app/boards/BestArticleList.tsx
@@ -8,7 +8,7 @@ import ArticleSkeleton from './ArticleSkeleton';
  * @param {string | undefined} props.keyword - 입력 검색어
  * @returns {JSX.Element} 베스트 게시글 리스트 컴포넌트
  */
-function BestArticleList({ keyword }: { keyword: string | undefined }) {
+function BestArticleList() {
   const deviceType = useDeviceType();
   let pageSize = 3;
 
@@ -32,33 +32,27 @@ function BestArticleList({ keyword }: { keyword: string | undefined }) {
   if (isError) return '에러가 발생했습니다.';
 
   return (
-    <>
-      {!keyword ? (
-        <section className="mt-pr-40 border border-x-0 border-t-0 pb-pr-40 mo:mt-pr-24 mo:pb-pr-32">
-          <h3 className="text-20b">베스트 게시글</h3>
+    <section className="mt-pr-40 border border-x-0 border-t-0 pb-pr-40 mo:mt-pr-24 mo:pb-pr-32">
+      <h3 className="text-20b">베스트 게시글</h3>
 
-          <div className="mt-pr-32 flex gap-x-pr-20 mo:mt-pr-24 ta:gap-pr-16">
-            {!isLoading ? (
-              <>
-                {bestArticleList?.list.map((bestArticle) => {
-                  return (
-                    <ArticleCard
-                      key={bestArticle.id}
-                      type="best"
-                      articleData={bestArticle}
-                    />
-                  );
-                })}
-              </>
-            ) : (
-              <ArticleSkeleton type="best" count={pageSize} />
-            )}
-          </div>
-        </section>
-      ) : (
-        ''
-      )}
-    </>
+      <div className="mt-pr-32 flex gap-x-pr-20 mo:mt-pr-24 ta:gap-pr-16">
+        {!isLoading ? (
+          <>
+            {bestArticleList?.list.map((bestArticle) => {
+              return (
+                <ArticleCard
+                  key={bestArticle.id}
+                  type="best"
+                  articleData={bestArticle}
+                />
+              );
+            })}
+          </>
+        ) : (
+          <ArticleSkeleton type="best" count={pageSize} />
+        )}
+      </div>
+    </section>
   );
 }
