+++ app/boards/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { ChangeEvent, FormEvent, useState } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { useSearchParams } from 'next/navigation';
 
 import Container from '@/components/layout/Container';
 
@@ -14,34 +13,15 @@ import ArticleSearch from './ArticleSearch';
  * @returns {JSX.Element} 자유게시판 페이지
  */
 export default function BoardsPage() {
-  const [search, setSearch] = useState('');
   const searchParams = useSearchParams();
   const keyword = searchParams.get('q') ?? undefined;
 
-  const router = useRouter();
-
-  const handleSearchInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setSearch(e.target.value);
-  };
-
-  const handleSearchFormSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    if (search) {
-      router.push(`/boards?q=${search}`);
-    } else if (search === '') {
-      router.push('/boards');
-    }
-  };
-
   return (
     <Container className="relative pb-pr-40">
-      <ArticleSearch
-        search={search}
-        handleSearchInputChange={handleSearchInputChange}
-        handleSearchFormSubmit={handleSearchFormSubmit}
-      />
-      <BestArticleList keyword={keyword} />
+      <ArticleSearch />
+
+      {!keyword && <BestArticleList />}
+
       <ArticleList keyword={keyword} />
       {!keyword && <AddArticleButton />}
     </Container>
